Changed the name of run_gw.distance_matrix_preload_global to compute_GW_distance_preload_global. (This shows it computes the GW distance matrix, not the intracellular idstance matrix)
Fixed error in documentation of the aforementioned function.

Renamed save_dist_mat_preload_global to compute_and_save_GW_dist_mat, as the new name makes it clear that
1. it does much more than just perform write-to-file operations, it actually calls the GW distance matrix computation
2. the name GW should be in the title to reflect that this is GW distance not intracellular distance

The docstring for the aforemention has also been updated to reflect that it computes the GW distance.

Notebook tutorials updated appropriately.