Removed function argument to plot_networkx
The latest version of networkx seems to have rearranged things so that labelling is handled by a different function now. The argument "with_labels=False" gives me an error "with_labels argument not recognized". The documentation for the current version of networkx does not mention with_labels as an argument of this function.